correção do item Sem Pagamento
