+++ lib/screens/viagens/all_viagens.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:rodagem/models/detail_viagens.dart';
+import 'package:rodagem/models/register_viagem.dart';
 import 'package:rodagem/screens/viagens/detail_viagem.dart';
 import 'package:rodagem/widget/item_viagem.dart';
 
@@ -127,7 +127,7 @@ class _AllViagensState extends State<AllViagens> {
                           itemBuilder: (_, indice) {
                             List<DocumentSnapshot> viagens = querySnapshot.documents.toList();
                             DocumentSnapshot documentSnapshot = viagens[indice];
-                            DetailViagens registerViagem = DetailViagens.fromDocumentSnapshot(documentSnapshot);
+                            RegisterViagem registerViagem = RegisterViagem.fromDocumentSnapshot(documentSnapshot);
                               return ItemViagens(
                                 viagens: registerViagem,
                                 onTapItem: () {
